+++ test/node.test.ts
@@ -1,0 +1,4 @@
+import test from '@playwright/test';
+import '../src/exdom';
+
+test('Make sure exdom can be imported in Node.js', () => {});