+++ scripts/test-coverage.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 set -o errexit
+set -o nounset
 set -o pipefail
 set -x
 
@@ -7,4 +8,4 @@ rm -rf coverage test-build
 
 tsc --project tsconfig.test.json
 
-playwright test $1 $2 $3
+c8 --include src --include test-build/src --reporter text --reporter lcov playwright test