+++ test/node.test.ts
@@ -1,4 +1,4 @@
-import test from '@playwright/test';
-import '../src/exdom';
+import { test } from '@playwright/test';
+import '../src/exdom.js';
 
 test('Make sure exdom can be imported in Node.js', () => {});